Addressing some sync issues by better implementation of setPrepared(). #61 Adding percentages to Sync. Fixing #60 sync damaging Addressing some sync issues Adding option to delete all Added Rsync and refactored some things

+++ shared-ui/src/main/java/es/uni_freiburg/de/cmotion/shared_ui/SRTHelper.java
@@ -16,7 +16,7 @@ import es.uni_freiburg.de.cmotion.shared_ui.model.SensorModel;
 
 public class SRTHelper {
 
-    public static double sRecordingDurationSec = -1;
+    public static long sRecordingDurationSec = -1;
 
 
     public static void sendRecordIntent(final Context context, List<SensorModel> selectedList) {
